refac: backend code 리팩토링, 파일 구조 정리
