+++ config/database.js
@@ -12,4 +12,4 @@ const connectDB = async () => {
   }
 };
 
-connectDB();
+module.exports = connectDB;
