+++ error/index.js
@@ -1,0 +1,17 @@
+const createError = require("http-errors");
+
+const errorHandler = (app) => {
+  app.use(function (req, res, next) {
+    next(createError(404));
+  });
+
+  app.use(function (err, req, res, next) {
+    res.locals.message = err.message;
+    res.locals.error = req.app.get("env") === "development" ? err : {};
+
+    res.status(err.status || 500);
+    res.render("error");
+  });
+};
+
+module.exports = errorHandler;
